x/twitter/types/message_add_tweet_test.go
package types

import (
	"testing"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
	"github/venkat/twitter/testutil/sample"
)

func TestMsgAddTweet_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgAddTweet
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgAddTweet{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgAddTweet{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
